Add: general election data collection, printing
Added general data collection about the election, such as when
the election happened, and if the election was a tie. Currently,
the data gets printed to standard out after being formatted nicely
as a JSON object.

Several functions were added: postListeners, tallyVotes,
averageVotes, checkTied, checkRandom, and requestInfo.

The individual song information is currently not included, and
prints out as an empty array.