Add: fiddling with restricted data
Added some code to the data parser to make it restrict data down to
the difference between the song's rating and the song with the highest
rating in the election, and whether or not the song was requested or not.
This lead to some interesting issues with KNN, so I tried KNN
regression; it didn't work as well. Yet.